change due to move card comp
