+++ src/components/Dashboard/CustomTabs/CustomTabs.tsx
@@ -9,7 +9,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 // core components
-import Card from '../Card/Card';
+import { Card } from '../../common';
 import CardBody from '../Card/CardBody';
 import CardHeader from '../Card/CardHeader';
