+++ src/components/Dashboard/views/Notifications/Notifications.tsx
@@ -12,7 +12,7 @@ import GridContainer from '../../Grid/GridContainer';
 import Button from '../../CustomButtons/Button';
 import SnackbarContent from '../../Snackbar/SnackbarContent';
 import Snackbar from '../../Snackbar/Snackbar';
-import Card from '../../Card/Card';
+import { Card } from '../../../common/Card/Card';
 import CardHeader from '../../Card/CardHeader';
 import CardBody from '../../Card/CardBody';
