+++ src/components/Dashboard/views/RTLPage/RTLPage.tsx
@@ -25,7 +25,7 @@ import Table from '../../Table/Table';
 import Tasks from '../../Tasks/Tasks';
 import CustomTabs from '../../CustomTabs/CustomTabs';
 import Danger from '../../Typography/Danger';
-import Card from '../../Card/Card';
+import { Card } from '../../../common/Card/Card';
 import CardHeader from '../../Card/CardHeader';
 import { CardAvatar } from '../../Card/CardAvatar';
 import CardIcon from '../../Card/CardIcon';
